+++ .gitignore
@@ -1,4 +1,6 @@
 node_modules
 dist
 examples/playground
+examples/package.json
+examples/yarn.lock
 .idea
